+++ o8g/Scripts/autoscripts.py
@@ -282,7 +282,7 @@ def useAbility(card, x = 0, y = 0): # The start of autoscript activation.
       return
    debugNotify("Not a tracing card. Checking highlight...", 4)
    if len(hijackDefaultAction): 
-      hijcack(card)
+      hijack(card)
       return
    if markerScripts(card): return # If there's a special marker, it means the card triggers to do something else with the default action
    if card.highlight == InactiveColor: