+++ o8g/Scripts/constants.py
@@ -69,6 +69,7 @@ specialHostPlacementAlgs = { # A Dictionary which holds tuples of X and Y placem
                               'The Supplier'          :            (-32,0),
                               'Awakening Center'      :            (-32,0),
                               'London Library'        :            (-32,0),
+                              'Street Peddler'        :            (-32,0),
                               'Off-Campus Apartment'  :            (-32,0)
                            }